# Implement intersperse using SSE2