@Bodigrim 2020-10-24
Looks reasonable to me, but I have very limited knowledge of SSE intrinsics. @vdukhovni @0xd34df00d could you possibly help me out here?