@Bodigrim 2020-10-24
> I can make a PR configuring GitHub Actions to build and test bytestring on Linux, Mac and Windows if you're OK with that.

@ethercrow  This would be a much appreciated contribution indeed.