@Bodigrim 2020-10-25
> I just want the concept of using SIMD in Haskell to be more mainstream.

I'm sympathetic to this goal. While ideally SIMD should be GHC primops, this will not happen without more demand from the user space. And this demand is to grow from experiments and explorations like this.