@ethercrow 2020-11-04
Added a benchmark on unaligned data, performance is the same:

```
benchmarked intersperse/intersperse
time                 626.8 ns   (624.0 ns .. 631.5 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 632.2 ns   (630.6 ns .. 633.9 ns)
std dev              5.852 ns   (4.659 ns .. 7.552 ns)

benchmarked intersperse/intersperse (unaligned)
time                 628.3 ns   (623.8 ns .. 633.6 ns)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 635.3 ns   (632.3 ns .. 640.8 ns)
std dev              13.55 ns   (9.235 ns .. 23.72 ns)
```