@ethercrow 2020-11-04
Windows build failed, but not because of the intrinsics, I think:

```
C:\Users\RUNNER~1\AppData\Local\Temp\ghcF6D5.o: DeleteFile "\\\\?\\C:\\Users\\RUNNER~1\\AppData\\Local\\Temp\\ghcF6D5.o": permission denied (Access is denied.)
```