@ethercrow 2020-11-07
@Bodigrim rebased onto your windows fixes in master, now looks fine.