@Bodigrim 2020-11-07
I do not have particular reservations here, because even I can read this code and understand what's going on and probably even fix something trivial. And I believe that userland package should pave a way for intrinsics in GHC, so I'm OK with additional complexity for a not-so-important function. But I do not have expertise to judge about crossplatform and crossarchitecture issues; this is where I would like to hear from @vdukhovni. 

Starting from GHC 8.10.2 FreeBSD is a Tier 1 platform. AFAIK there are no CIs available, but one can create a FreeBSD droplet on DigitalOcean to run tests and destroy it in an hour. Could someone conjure up a shell script for FreeBSD 12.2 zfs x64 to install ghc / cabal and run `bytestring` tests?