@vdukhovni 2020-11-07
> Starting from GHC 8.10.2 FreeBSD is a Tier 1 platform. AFAIK there are no CIs available, but one can create a FreeBSD droplet on DigitalOcean to run tests and destroy it in an hour. Could someone conjure up a shell script for FreeBSD 12.2 zfs x64 to install ghc / cabal and run `bytestring` tests?

On FreeBSD 12, ghc and cabal-install are available via the "ports" system.
```
$ pkg search ghc
ghc-8.10.2                     Compiler for the functional language Haskell
$ pkg search cabal
hs-cabal-install-3.2.0.0       Command-line interface for Cabal and Hackage
```
A job running as "root" can install ports via: "pkg install ghc cabal-install".  After that it is not different from what one would do on Linux.  I have no experience with "droplets". Are they able to install ports?