@Bodigrim 2020-11-07
"Droplet" is just a fancy name for a virtual machine. Thanks, this works:
```
pkg install git ghc hs-cabal-install
cabal update
git clone https://github.com/haskell/bytestring.git
cd bytestring
cabal build 
(cd tests; cabal test)
(cd bench; cabal bench -O0 --benchmark-options "--quick --min-duration=0 --include-first-iter")
```
The branch seems working fine as well.