@Bodigrim 2020-11-07
> [you meant to say Tier 2]

Why? There is an official binary build: https://www.haskell.org/ghc/download_ghc_8_10_2.html#freebsd_x86_64