@vdukhovni 2020-11-07
> > [you meant to say Tier 2]
> 
> Why? There is an official binary build: https://www.haskell.org/ghc/download_ghc_8_10_2.html#freebsd_x86_64

Perhaps the Wiki page I found is [outdated](
https://gitlab.haskell.org/ghc/ghc/-/wikis/platforms#:~:text=Criteria%20for%20Tier%201%20platforms,us%20during%20the%20release%20process)