@ethercrow 2020-11-12
This should take care of testing on FreeBSD: https://github.com/haskell/bytestring/pull/322