@vdukhovni 2020-11-19
> @vdukhovni is anything outstanding left here?

I have no outstanding issues.  The motivation is a bit weak for this particular instance, as I mentioned at the outset, but if the goal is to start with the simplest (even if not compelling) applications and build from there, than it makes a bit of sense.